requirements-dev.txt
django-test-plus==1.4.0
factory-boy==2.12.0
flake8==3.8.3
pip-tools
pytest-django==3.9.0
